+++ src/main/java/com/boardgo/domain/mapper/UserInfoMapper.java
@@ -1,12 +1,12 @@
 package com.boardgo.domain.mapper;
 
-import com.boardgo.domain.user.controller.dto.OtherPersonalInfoResponse;
-import com.boardgo.domain.user.controller.dto.SignupRequest;
-import com.boardgo.domain.user.controller.dto.UserPersonalInfoResponse;
+import com.boardgo.domain.user.controller.request.SignupRequest;
 import com.boardgo.domain.user.entity.UserInfoEntity;
+import com.boardgo.domain.user.repository.projection.PersonalInfoProjection;
 import com.boardgo.domain.user.repository.projection.UserParticipantProjection;
-import com.boardgo.domain.user.repository.response.PersonalInfoDto;
 import com.boardgo.domain.user.repository.response.UserParticipantResponse;
+import com.boardgo.domain.user.service.response.OtherPersonalInfoResponse;
+import com.boardgo.domain.user.service.response.UserPersonalInfoResponse;
 import com.boardgo.oauth2.dto.OAuth2CreateUserRequest;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -25,7 +25,7 @@ public interface UserInfoMapper {
     UserInfoEntity toUserInfoEntity(OAuth2CreateUserRequest oAuth2CreateUserRequest);
 
     UserPersonalInfoResponse toUserPersonalInfoResponse(
-            PersonalInfoDto userPersonalInfoResponse, Double averageRating);
+            PersonalInfoProjection userPersonalInfoResponse, Double averageRating);
 
     UserParticipantResponse toUserParticipantResponse(
             UserParticipantProjection userParticipantProjection);
