+++ src/main/java/com/boardgo/domain/user/controller/request/EmailRequest.java
@@ -1,4 +1,4 @@
-package com.boardgo.domain.user.controller.dto;
+package com.boardgo.domain.user.controller.request;
 
 import jakarta.validation.constraints.Email;
