+++ src/main/java/com/boardgo/domain/user/controller/request/UserPersonalInfoUpdateRequest.java
@@ -1,3 +1,3 @@
-package com.boardgo.domain.user.controller.dto;
+package com.boardgo.domain.user.controller.request;
 
 public record UserPersonalInfoUpdateRequest(String nickName, String password) {}
