+++ src/main/java/com/boardgo/domain/user/repository/projection/PersonalInfoProjection.java
@@ -1,10 +1,10 @@
-package com.boardgo.domain.user.repository.response;
+package com.boardgo.domain.user.repository.projection;
 
 import com.querydsl.core.annotations.QueryProjection;
 import java.util.List;
 
-public record PersonalInfoDto(
+public record PersonalInfoProjection(
         String email, String nickName, String profileImage, List<String> prTags) {
     @QueryProjection
-    public PersonalInfoDto {}
+    public PersonalInfoProjection {}
 }
