+++ src/main/java/com/boardgo/domain/user/service/UserCommandServiceV1.java
@@ -11,9 +11,9 @@ import com.boardgo.common.exception.DuplicateException;
 import com.boardgo.common.utils.FileUtils;
 import com.boardgo.common.utils.S3Service;
 import com.boardgo.domain.mapper.UserInfoMapper;
-import com.boardgo.domain.user.controller.dto.SignupRequest;
-import com.boardgo.domain.user.controller.dto.SocialSignupRequest;
-import com.boardgo.domain.user.controller.dto.UserPersonalInfoUpdateRequest;
+import com.boardgo.domain.user.controller.request.SignupRequest;
+import com.boardgo.domain.user.controller.request.SocialSignupRequest;
+import com.boardgo.domain.user.controller.request.UserPersonalInfoUpdateRequest;
 import com.boardgo.domain.user.entity.UserInfoEntity;
 import com.boardgo.domain.user.entity.UserPrTagEntity;
 import com.boardgo.domain.user.repository.UserPrTagRepository;
