+++ src/main/java/com/boardgo/jwt/JWTFilter.java
@@ -1,9 +1,10 @@
 package com.boardgo.jwt;
 
-import static com.boardgo.common.constant.HeaderConstant.*;
+import static com.boardgo.common.constant.HeaderConstant.AUTHORIZATION;
+import static com.boardgo.common.constant.HeaderConstant.BEARER;
 
 import com.boardgo.domain.user.entity.UserInfoEntity;
-import com.boardgo.domain.user.service.dto.CustomUserDetails;
+import com.boardgo.domain.user.service.response.CustomUserDetails;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
