+++ src/test/java/com/boardgo/integration/user/service/UserQueryServiceV1Test.java
@@ -11,17 +11,17 @@ import com.boardgo.common.exception.CustomIllegalArgumentException;
 import com.boardgo.common.exception.CustomNullPointException;
 import com.boardgo.domain.meeting.repository.MeetingParticipantRepository;
 import com.boardgo.domain.review.repository.ReviewRepository;
-import com.boardgo.domain.user.controller.dto.EmailRequest;
-import com.boardgo.domain.user.controller.dto.NickNameRequest;
-import com.boardgo.domain.user.controller.dto.OtherPersonalInfoResponse;
-import com.boardgo.domain.user.controller.dto.UserPersonalInfoResponse;
+import com.boardgo.domain.user.controller.request.EmailRequest;
+import com.boardgo.domain.user.controller.request.NickNameRequest;
 import com.boardgo.domain.user.entity.UserInfoEntity;
 import com.boardgo.domain.user.entity.UserPrTagEntity;
 import com.boardgo.domain.user.entity.enums.ProviderType;
 import com.boardgo.domain.user.repository.UserPrTagRepository;
 import com.boardgo.domain.user.repository.UserRepository;
 import com.boardgo.domain.user.service.UserQueryUseCase;
-import com.boardgo.domain.user.service.dto.CustomUserDetails;
+import com.boardgo.domain.user.service.response.CustomUserDetails;
+import com.boardgo.domain.user.service.response.OtherPersonalInfoResponse;
+import com.boardgo.domain.user.service.response.UserPersonalInfoResponse;
 import com.boardgo.integration.init.TestUserInfoInitializer;
 import com.boardgo.integration.support.IntegrationTestSupport;
 import java.util.List;
